Added main form, fixes #2
