+++ mainwindow.h
@@ -2,13 +2,7 @@
 #define MAINWINDOW_H
 
 #include <QMainWindow>
-#include <QGraphicsView>
-#include <map>
-#include "displayabledfdelement.h"
-#include "directedgraph.h"
-#include "dfdelement.h"
-#include <QDate>
-#include "clickableellipseitem.h"
+#include "planwindow.h"
 
 namespace Ui {
 class MainWindow;
@@ -21,12 +15,12 @@ class MainWindow : public QMainWindow
 public:
     explicit MainWindow(QWidget *parent = 0);
     ~MainWindow();
-    void drawDFD(bool rebuildMap);
+
+private slots:
+    void on_createPlanPushButton_clicked();
 
 private:
     Ui::MainWindow *ui;
-    DirectedGraph<DFDElement> *graph;
-    std::map <int, DisplayableDFDElement*> elMap;
 };
 
 #endif // MAINWINDOW_H
